core/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.schemas import get_schema_view
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
)

urlpatterns = [
    path('api/token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('api/token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('admin/', admin.site.urls),
    path('', include('uppy.urls', namespace='uppy')),
    path('api/', include('uppy_api.urls', namespace='uppy_api')),
    path('api/user/', include('users.urls', namespace='users')),
    path('api-auth/', include('rest_framework.urls', namespace='rest_framework')),
    path('schema/', get_schema_view(
        title="Uppy",
        description="API for Uppy",
        version="1.0.0"
    ), name='openapi-schema'),
]
